requirements.txt
fastapi
uvicorn
phidata
pgvector
pypdf
psycopg[binary]
sqlalchemy
phidata[aws]
openai